Use size_t (instead of int) whereever the manpage of the respective function specifies size_t.
Acked-by: Peter Stuge <peter@stuge.se>



git-svn-id: svn://coreboot.org/repository/LinuxBIOSv3@301 f3766cd6-281f-0410-b1cd-43a5c92072e9

+++ include/string.h
@@ -24,10 +24,10 @@
 #include <arch/types.h>
 
 /* Prototypes for functions from lib/mem.c. */
-extern void *memcpy(void *dest, const void *src, int len);
-extern void *memmove(void *dest, const void *src, int len);
-extern void *memset(void *v, unsigned char a, int len);
-extern int memcmp(const void *s1, const void *s2, int len);
+extern void *memcpy(void *dest, const void *src, size_t len);
+extern void *memmove(void *dest, const void *src, size_t len);
+extern void *memset(void *v, unsigned char a, size_t len);
+extern int memcmp(const void *s1, const void *s2, size_t len);
 
 /* Prototypes for functions from console/vsprintf.c. */
 extern int sprintf(char *buf, const char *fmt, ...);
